make backups not poll() too long
This fixes a situation where there are multiple backups as well as
multiple masters (on different VHIDs). A packet received from a
different master will interrupt the poll(), so poll()ing again, for the
same duration, will result in sleeping too long.

This is a better fix than my previous "also use advskew when not
advertising".